requirements.txt
easyocr
Pillow
gradio
pdf2image
